// src/modules/enums.ts
export enum Power {
  Fox = "Fox",
  FreeAction = "FreeAction",
  BlueCheck = "BlueCheck",
  GreenCheck = "GreenCheck",
  GreyCheck = "GreyCheck",
  PinkCheck = "PinkCheck",
  PurpleSix = "PurpleSix",
  YellowCheck = "YellowCheck",
  ReRoll = "ReRoll",
  FitNumber = "FitNumber",
  OrangeFour = "OrangeFour",
  OrangeFive = "OrangeFive",
  OrangeSix = "OrangeSix",
  Double = "Double",
  Triple = "Triple"
}